Fix loss of data due to conversion warnings in MS-Windows

+++ src/filepath.c
@@ -675,7 +675,7 @@ repeat:
 			    if (s != NULL)
 			    {
 				*fnamep = s;
-				*fnamelen = slen;
+				*fnamelen = (int)slen;
 				vim_free(*bufp);
 				*bufp = s;
 				didit = TRUE;
